scenario_maker/requirements.txt
-e git://github.com/AndreaCensi/zuper_utils.git@z1#egg=zuper_utils
-e git://github.com/AndreaCensi/zuper_nodes.git@z1#egg=zuper_nodes
-e git://github.com/duckietown/aido-protocols.git@v4#egg=aido-protocols

numpy==1.16.2

PyGeometry


-e git://github.com/duckietown/duckietown-world.git@aido2_lf_r1#egg=duckietown-world

